requirements.txt
transformers>=4.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
gradio>=4.0.0
torch>=2.0.0
numpy>=1.23.0
regex>=2023.0.0
